Flask-DB-Project/requirements.txt
astroid==2.4.2
click==7.1.2
colorama==0.4.3
Flask==1.1.2
flask-ldap3-login==0.9.16
Flask-Login==0.5.0
Flask-SQLAlchemy==2.4.4
Flask-WTF==0.14.3
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.11.2
lazy-object-proxy==1.4.3
ldap3==2.7
MarkupSafe==1.1.1
mccabe==0.6.1
numpy==1.19.1
pandas==1.1.0
pyasn1==0.4.8
pylint==2.5.3
PyMySQL==0.10.0
pyodbc==4.0.30
pypyodbc==1.3.4
python-dateutil==2.8.1
pytz==2020.1
PyYAML==5.3.1
six==1.15.0
SQLAlchemy==1.3.18
toml==0.10.1
typed-ast==1.4.1
Werkzeug==1.0.1
wrapt==1.12.1
WTForms==2.3.3
